Moved sieve ordering to Assembler object
