+++ src/main/scala/org/clulab/assembly/Assembler.scala
@@ -45,7 +45,7 @@ class Assembler(mns: Seq[Mention]) extends LazyLogging {
     * For the given parent event mention, find the features for the event's participants
     * and return them in a map, keyed by each participant of the given parent event.
     */
-  def getInputFeaturesByParticipants (parent: Mention): Map[Mention,RoleWithFeatures] = {
+  def getInputFeaturesByParticipants(parent: Mention): Map[Mention,RoleWithFeatures] = {
     val features = getInputFeaturesForParticipants(parent)
     features.map(rwfs => rwfs.participant -> rwfs).toMap
   }
@@ -90,4 +90,59 @@ class Assembler(mns: Seq[Mention]) extends LazyLogging {
   def getEquivalenceLinks(m: Mention): Set[Equivalence] =
     equivalenceLinks.getOrElse(m, Set.empty[Equivalence])
 
+}
+
+
+object Assembler extends LazyLogging {
+
+  /**
+    * Applies Assembly Sieves to mentions and returns and updated AssemblyManager.
+    *
+    * @param mentions a Seq of Odin Mentions
+    * @return an AssemblyManager
+    */
+  def applySieves(mentions: Seq[Mention]): AssemblyManager = {
+
+    val dedup = new DeduplicationSieves()
+    val precedence = new PrecedenceSieves()
+
+    val orderedSieves =
+    // track relevant mentions
+      AssemblySieve(dedup.trackMentions) andThen
+        // find precedence relations using rules
+        AssemblySieve(precedence.withinRbPrecedence) andThen
+        AssemblySieve(precedence.reichenbachPrecedence) andThen
+        AssemblySieve(precedence.betweenRbPrecedence) andThen
+        AssemblySieve(precedence.featureBasedClassifier)
+
+    // apply the sieves and return the manager
+    val am: AssemblyManager = orderedSieves.apply(mentions)
+
+    am
+  }
+
+  /**
+    * Applies each Assembly Sieve to mentions and returns and updated AssemblyManager for each.
+    *
+    * @param mentions a Seq of Odin Mentions
+    * @return an AssemblyManager
+    */
+  def applyEachSieve(mentions: Seq[Mention]): Map[String, AssemblyManager] = {
+
+    val dedup = new DeduplicationSieves()
+    val precedence = new PrecedenceSieves()
+
+    val availableSieves = Map(
+      "withinRbPrecedence" -> (AssemblySieve(dedup.trackMentions) andThen AssemblySieve(precedence.withinRbPrecedence)),
+      "reichenbachPrecedence" -> (AssemblySieve(dedup.trackMentions) andThen AssemblySieve(precedence.reichenbachPrecedence)),
+      "betweenRbPrecedence" -> (AssemblySieve(dedup.trackMentions) andThen AssemblySieve(precedence.betweenRbPrecedence))
+    )
+
+    val ams = for {
+      (lbl, s) <- availableSieves.par
+      am = s.apply(mentions)
+    } yield lbl -> am
+
+    ams.seq //++ Map("all" -> applySieves(mentions))
+  }
 }
